14536: JazzyNico 2015-12-15 Tutorials. German and Greek translations update.

+++ inkscape-launchpad/src/style-internal.h
@@ -342,7 +342,7 @@ public:
           computed(0)
     {}
 
-    SPILength( Glib::ustring const &name, unsigned value = 0 )
+    SPILength( Glib::ustring const &name, float value = 0 )
         : SPIBase( name ),
           unit(SP_CSS_UNIT_NONE),
           value(value),
@@ -401,7 +401,7 @@ public:
           normal(true)
     {}
 
-    SPILengthOrNormal( Glib::ustring const &name, unsigned value = 0 )
+    SPILengthOrNormal( Glib::ustring const &name, float value = 0 )
         : SPILength( name, value ),
           normal(true)
     {}
